backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
httpx
python-dotenv
sentence-transformers
chromadb
anthropic
ollama
Pillow
